updated: notes of reviewer /n refactor index.js /n refactor below js files to /src/components directory Popup.js PopupWithForm.js Sections.js UserInfo.js
